refactor: Move MockedResponse class to separate file
