+++ mopinion/test/mocks/__init__.py
@@ -1,0 +1,1 @@
+from .mock_response import MockedResponse
